New registry related fixes

+++ local-tap.sh
@@ -100,21 +100,46 @@ while test $# -gt 0; do
       insecure_expose_kube_api=$1
       shift
       ;;
+    --install-from-local-registry)
+      shift
+      install_from_local_registry=$1
+      shift
+      ;;
+    --target)
+      shift
+      target=$1
+      shift
+      ;;
+    --keep-local-registry)
+      shift
+      keep_local_registry=$1
+      shift
+      ;;
+    --cluster-name)
+      shift
+      cluster_name=$1
+      shift
+      ;;
     --help)
       cat << EOF
 Usage: local-tap.sh [OPTIONS]
 Options:
 
 [Global Manadatory Flags]
-  --action : What action to take - create,stop,start,status,delete,prepare
+  --action : What action to take - create,stop,start,status,delete,prepare,relocate-images
 
 [Global Optional Flags]
   --help : show this help menu
+  --cluster-name : This allows you to set the cluster name to work against. (Default: tce-tap)
 
 [Mandatory Flags - For Create Action]
   --tanzunet-user : User for Tanzu Network
   --tanzunet-password : Password for Tanzu Network
   --tap-package-repo-url : URL For Relocated TAP Package Repository
+
+[Mandatory Flags - For Relocate Images Action]
+  --tanzunet-user : User for Tanzu Network
+  --tanzunet-password : Password for Tanzu Network
 
 [Optional Flags - For Create Action]
   --tap-profile : TAP Installation Profile. (Default full)
@@ -132,6 +157,13 @@ Options:
   --enable-remote-access : (yes or no) This flag allows you to set whether remote access from other machines should be allowed to TAP GUI and other exposed endpoints (Default: no)
   --ip-address : Required if enabling remote access. This flags value needs to be the IP of your node (eg. 192.168.1.231)
   --insecure-expose-kube-api : (yes or no) This flag allows you to set whether the Kubernetes Cluster will be configured to allow access from outside your machine. (Default: no)
+  --install-from-local-registry : (yes or no) If you relocated the TAP images to a local docker registry on your machine set to yes (Default: no)
+
+[Optional Flags - For Relocate Images Action]
+  --target : (local or remote) set to local if pushing to a local docker registry otherwise set to remote (Default: remote)
+
+[Optional Flags - For Delete Action]
+  --keep-local-registry (yes or no) If set to yes, the local registry will not be deleted. (Default: no)
 
 EOF
       exit 1
@@ -167,6 +199,9 @@ BR='\033[1;31m'         # Red
 BG='\033[1;32m'       # Green
 BY='\033[1;33m'      # Yellow
 B='\033[1m'                   # Bold Regular
+RNC=$'\033[0m'
+RB=$'\033[1m'
+BU='\033[1m\e[4m'
 
 print_package_status () {
   pkg_status=`kubectl --context kind-tce-tap get pkgi -n $2 $3 -o custom-columns=STATUS:.status.friendlyDescription --no-headers`
@@ -181,7 +216,134 @@ print_package_status () {
   fi
 }
 start=`date +%s`
-if [[ $action == "prepare" ]]; then
+if ! [[ $cluster_name ]]; then
+  cluster_name="tce-tap"
+fi
+if [[ $action == "relocate-images" ]]; then
+  clear
+  echo -e "${BU}Prereq Check:${NC}"
+  echo ""
+  fail="no"
+  if ! [[ $target ]]; then
+    target="remote"
+  fi
+  if ! command -v docker &> /dev/null
+  then
+    echo -e "${BR}ERROR: Docker is not installed. please install docker and try again.${NC}"
+    fail="yes"
+  else
+    echo -e "${BG}Docker is installed${NC}"
+  fi
+  if ! command -v imgpkg &> /dev/null
+  then
+    echo -e "${BR}ERROR: imgpkg is not installed. please install imgpkg and try again.${NC}"
+    fail="yes"
+  else
+    echo -e "${BG}imgpkg is installed${NC}"
+  fi
+  if [[ $fail == "yes" ]]; then
+    echo -e "${BR}Prereq check failed. Please install the prerequisite tools (docker and imgpkg) and then try again${NC}"
+    exit 1
+  fi
+  get_password() {
+    unset PWORD
+    PWORD=
+    echo -e -n "${B}$1${NC}" 1>&2
+    while IFS= read -r -n1 -s char; do
+      # Convert users key press to hexadecimal character code
+      # Note a 'return' or EOL, will return a empty string
+      #
+      #code=$( echo -n "$char" | od -An -tx1 | tr -d ' \011' )
+      code=${char:+$(printf '%02x' "'$char'")} # set to nothing for EOL
+  
+      case "$code" in
+      ''|0a|0d) break ;;   # EOL, newline, return
+      08|7f)  # backspace or delete
+          if [ -n "$PWORD" ]; then
+            PWORD="$( echo "$PWORD" | sed 's/.$//' )"
+            echo -n $'\b \b' 1>&2
+          fi
+          ;;
+      15) # ^U or kill line
+          echo -n "$PWORD" | sed 's/./\cH \cH/g' >&2
+          PWORD=''
+          ;;
+      [01]?) ;; # Ignore ALL other control characters
+      *)  PWORD="$PWORD$char"
+          echo -n '*' 1>&2
+          ;;
+      esac
+    done
+    echo
+    echo $PWORD
+  }
+  echo ""
+  echo -e "${BU}TAP Package Repository Relocation Info:${NC}"
+  echo ""
+  echo -e "${BU}Tanzu Network Details${NC}"
+  echo ""
+  read -p "${RB}Tanzu Network Username: ${RNC}" TANZUNET_USER
+  
+  # read -s -p 'Tanzu Network Password: ' TANZUNET_PASSWORD
+  TANZUNET_PASSWORD="$(get_password 'Tanzu Network Password: ')"
+  echo ""
+  TANZUNET_PASSWORD_CONFIRM="$(get_password 'Tanzu Network Password (Agian): ')"
+  echo ""
+  while [ "$TANZUNET_PASSWORD" != "$TANZUNET_PASSWORD_CONFIRM" ];
+  do
+    echo -e "${BR}Passwords do not match. Please try again${NC}"
+    TANZUNET_PASSWORD="$(get_password 'Tanzu Network Password: ')"
+    echo ""
+    TANZUNET_PASSWORD_CONFIRM="$(get_password 'Tanzu Network Password (Agian): ')"
+    echo ""
+  done
+  if [[ $target == "local" ]]; then
+    echo "${RB}Deploying a local registry to hold TAP images${RNC}"
+    docker run -d --restart=always -p "5000:5000" --name "registry.local" registry:2
+    INSTALL_REGISTRY_REPO="tap"
+    INSTALL_REGISTRY_HOSTNAME="localhost:5000"
+    INSTALL_REGISTRY_PASSWORD="password"
+    INSTALL_REGISTRY_USERNAME="user"
+  else
+    echo ""
+    echo ""
+    echo -e "${BU}Destination Repository Details${NC}"
+    echo ""
+    read -p "${RB}Destination Registry FQDN: ${RNC}" INSTALL_REGISTRY_HOSTNAME
+    read -p "${RB}Destination Registry Repo: ${RNC}" INSTALL_REGISTRY_REPO
+    read -p "${RB}Destination Registry Username: ${RNC}" INSTALL_REGISTRY_USERNAME
+    
+    # read -s -p 'Destination Registry Password: ' INSTALL_REGISTRY_PASSWORD
+    INSTALL_REGISTRY_PASSWORD="$(get_password 'Destination Registry Password: ')"
+    echo ""
+    INSTALL_REGISTRY_PASSWORD_CONFIRM="$(get_password 'Destination Registry Password (Agian): ')"
+    echo ""
+    while [ "$INSTALL_REGISTRY_PASSWORD" != "$INSTALL_REGISTRY_PASSWORD_CONFIRM" ];
+    do
+      echo -e "${BR}Passwords do not match. Please try again${NC}"
+      INSTALL_REGISTRY_PASSWORD="$(get_password 'Destination Registry Password: ')"
+      echo ""
+      INSTALL_REGISTRY_PASSWORD_CONFIRM="$(get_password 'Destination Registry Password (Agian): ')"
+      echo ""
+    done
+  fi
+  read -p "${RB}TAP Version: ${RNC}" TAP_VERSION
+  echo ""
+  echo -e "${BU}The final URL for the package repository will be:${NC} ${INSTALL_REGISTRY_HOSTNAME}/${INSTALL_REGISTRY_REPO}/tap-packages:${TAP_VERSION}"
+  echo ""
+  read -p "${RB}Continue (y/n)? ${RNC}" choice
+  case "$choice" in
+    y|Y ) echo "Logging into the source and destination registries" ;;
+    n|N ) exit 1 ;;
+    * ) echo "invalid choice. exiting now." && exit 1;;
+  esac
+  echo ""
+  echo $TANZUNET_PASSWORD | docker login --username $TANZUNET_USER --password-stdin registry.tanzu.vmware.com
+  echo $INSTALL_REGISTRY_PASSWORD | docker login --username $INSTALL_REGISTRY_USERNAME --password-stdin $INSTALL_REGISTRY_HOSTNAME
+  echo ""
+  echo -e "${B}Relocating the images now. This may take a few minutes.${NC}"
+  imgpkg copy -b registry.tanzu.vmware.com/tanzu-application-platform/tap-packages:${TAP_VERSION} --to-repo ${INSTALL_REGISTRY_HOSTNAME}/${INSTALL_REGISTRY_REPO}/tap-packages
+elif [[ $action == "prepare" ]]; then
   echo "Installing Prereqs if not present or of the wrong version"
 
   if [[ $machine_os != "Mac" && $machine_os != "Linux" ]]; then
@@ -252,7 +414,7 @@ if [[ $action == "prepare" ]]; then
   fi
 
 elif [[ $action == "status" ]]; then
-  cls_status=`tanzu uc list -o json | jq -r '.[] | select(.name=="tce-tap") | .status'`
+  cls_status=`tanzu uc list -o json | jq -r '.[] | select(.name=="'$cluster_name'") | .status'`
   if [[ $cls_status == "Running" ]]; then
     echo -e "${B}Cluster Status: ${BG}$cls_status${NC}"
   elif [[ $cls_status == "Stopped" ]]; then
@@ -268,7 +430,7 @@ elif [[ $action == "status" ]]; then
   else
     echo -e "${B}Local Registry Status: ${BR}$reg_status${NC}"
   fi
-  if [[ `tanzu uc list -o json | jq -r '.[] | select(.name=="tce-tap") | .status'` == "Running" ]]; then
+  if [[ `tanzu uc list -o json | jq -r '.[] | select(.name=="'$cluster_name'") | .status'` == "Running" ]]; then
     echo -e "${B}Package Statuses:${NC}"
     tap_profile=`kubectl get secret -n tkg-system tap-config -o json | jq '.data["values.yml"]' -r | base64 -d | head -n 1 | cut -d ":" -f2`
     supply_chain_suffix=`kubectl get secret -n tkg-system tap-config -o json | jq -r '.data."values.yml"' | base64 -d | grep "supply_chain: " | sed 's/^supply_chain: //' | sed 's/_/-/g'`
@@ -308,7 +470,8 @@ elif [[ $action == "status" ]]; then
     fi
   fi
   echo -e "${B}Images in the Local Registry${NC}"
-  curl http://localhost:5000/v2/_catalog --silent | jq -r .repositories[] | sed 's/^/    /'
+  curl -s http://localhost:5000/v2/_catalog | jq -r .repositories[] |  while read repo ; do echo -e "    ${RB}$repo:${NC}" && imgpkg tag ls -i localhost:5000/$repo --json | jq -r .Tables[0].Rows[].name | sed 's/^/        /g' - ; done
+  #curl http://localhost:5000/v2/_catalog --silent | jq -r .repositories[] | sed 's/^/    /'
   end=`date +%s`
   runtime=$((end-start))
   hours=$((runtime / 3600))
@@ -317,11 +480,19 @@ elif [[ $action == "status" ]]; then
   echo ""
   echo "Script Runtime: $hours:$minutes:$seconds (hh:mm:ss)"
 elif [[ $action == "delete" ]]; then
-  echo "(1/2) Delete TCE TAP Cluster"
-  tanzu uc delete tce-tap
-  echo "(2/2) Delete Local Registry"
-  docker stop registry.local | sed 's/^/       /g'
-  docker rm registry.local | sed 's/^/       /g'
+  if ! [[ $keep_local_registry ]]; then
+    keep_local_registry="no"
+  fi
+  if [[ $keep_local_registry == "no" ]]; then
+    echo "(1/2) Delete TCE TAP Cluster"
+    tanzu uc delete $cluster_name
+    echo "(2/2) Delete Local Registry"
+    docker stop registry.local | sed 's/^/       /g'
+    docker rm registry.local | sed 's/^/       /g'
+  else
+    echo "(1/1) Delete TCE TAP Cluster"
+    tanzu uc delete $cluster_name
+  fi
   echo "Your environment has been deleted."
   end=`date +%s`
   runtime=$((end-start))
@@ -332,7 +503,7 @@ elif [[ $action == "delete" ]]; then
   echo "Script Runtime: $hours:$minutes:$seconds (hh:mm:ss)"
 elif [[ $action == "stop" ]]; then
   echo "(1/2) Stop TCE TAP Cluster"
-  tanzu uc stop tce-tap | sed 's/^/       /g'
+  tanzu uc stop $cluster_name | sed 's/^/       /g'
   echo "(2/2) Stop Local Registry"
   docker stop registry.local | sed 's/^/       /g'
   echo "Your environment has been stopped."
@@ -347,7 +518,7 @@ elif [[ $action == "start" ]]; then
   echo "(1/2) Start Local Registry"
   docker start registry.local | sed 's/^/       /g'
   echo "(2/2) Start TCE TAP Cluster"
-  tanzu uc start tce-tap | sed 's/^/       /g'
+  tanzu uc start $cluster_name | sed 's/^/       /g'
   echo "Your environment has been started. please give it a few minutes to come back up fully and reconcile all of the packages."
   end=`date +%s`
   runtime=$((end-start))
@@ -411,14 +582,17 @@ elif [[ $action == "create" ]]; then
   if ! [[ $ip_address ]]; then
     ip_address="127.0.0.1"
   fi
+  if ! [[ $install_from_local_registry ]]; then
+    install_from_local_registry="no"
+  fi
   task=1
-  mkdir -p tce-tap-files | sed 's/^/       /g'
-  cd tce-tap-files | sed 's/^/       /g'
+  mkdir -p $cluster_name-files | sed 's/^/       /g'
+  cd $cluster_name-files | sed 's/^/       /g'
   echo "($task/$task_count) Generating Config files"
   ((task++))
   if [[ $ca_file_path ]]; then
-    cat << EOF > tce-tap.yaml
-ClusterName: tce-tap
+    cat << EOF > $cluster_name.yaml
+ClusterName: $cluster_name
 Cni: calico
 PortsToForward:
   - HostPort: 80
@@ -427,8 +601,8 @@ PortsToForward:
     ContainerPort: 443
 EOF
   else
-    cat << EOF > tce-tap.yaml
-ClusterName: tce-tap
+    cat << EOF > $cluster_name.yaml
+ClusterName: $cluster_name
 Cni: calico
 AdditionalPackageRepos:
   - $tce_package_repo_url
@@ -450,7 +624,7 @@ InstallPackages:
 EOF
   fi
   if [[ $insecure_expose_kube_api == "yes" ]]; then
-    cat << EOF >> tce-tap.yaml
+    cat << EOF >> $cluster_name.yaml
 ProviderConfiguration:
   rawKindConfig: |
     networking:
@@ -608,7 +782,7 @@ EOF
   fi
   echo "($task/$task_count) Creating the TCE kind based unmanaged cluster"
   ((task++))
-  tanzu uc create -f tce-tap.yaml
+  tanzu uc create -f $cluster_name.yaml
   if [[ $ca_file_path ]]; then
     echo "($task/$task_count) Configuring Kapp Controller to Trust the provided CA"
     ((task++))
@@ -632,13 +806,18 @@ EOF
     tanzu package install -n tkg-system tap -p tap.tanzu.vmware.com -v $tap_version --wait=false -f tap-values.yaml
   fi
   kubectl create ns tap-install | sed 's/^/       /g'
-  echo "($task/$task_count) Creating a local docker registry to be used for TAP workloads and TBS images"
-  ((task++))
-  docker run -d --restart=always -p "5000:5000" --name "registry.local" registry:2 | sed 's/^/       /g'
+  if [[ $install_from_local_registry == "no" ]]; then
+    echo "($task/$task_count) Creating a local docker registry to be used for TAP workloads and TBS images"
+    ((task++))
+    docker run -d --restart=always -p "5000:5000" --name "registry.local" registry:2 | sed 's/^/       /g'
+  else
+    echo "($task/$task_count) Connecting your local docker registry to the TCE docker network to be used for TAP workloads and TBS images"
+    ((task++))
+  fi
   docker network connect kind registry.local | sed 's/^/       /g'
   echo "($task/$task_count) Configuring TCE cluster to trust the insecure local registry"
   ((task++))
-  docker cp tce-tap-control-plane:/etc/containerd/config.toml ./config.toml | sed 's/^/       /g'
+  docker cp $cluster_name-control-plane:/etc/containerd/config.toml ./config.toml | sed 's/^/       /g'
   tap_registry_fqdn=`echo $tap_package_repo_url | sed 's|/.*||g'`
   if [[ $dockerhub_registry_mirror ]]; then
     cat << EOF >> config.toml
@@ -671,8 +850,8 @@ EOF
       insecure_skip_verify = true
 EOF
   fi
-  docker cp config.toml tce-tap-control-plane:/etc/containerd/config.toml | sed 's/^/       /g'
-  docker exec tce-tap-control-plane service containerd restart | sed 's/^/       /g'
+  docker cp config.toml $cluster_name-control-plane:/etc/containerd/config.toml | sed 's/^/       /g'
+  docker exec $cluster_name-control-plane service containerd restart | sed 's/^/       /g'
   echo "($task/$task_count) Setting up Access to the local registry from your docker daemon..."
   ((task++))
   ipAddr=`docker inspect -f '{{.NetworkSettings.IPAddress}}' registry.local | tr '\n' ' '`
@@ -808,7 +987,7 @@ EOF
     kubectl rollout status deployment server -n tap-gui | sed 's/^/       /g'
   fi
   # Move config files to the dedicated folder
-  mv tce-tap.yaml tap-values.yaml tap-gui-secret-patch.yaml tap-gui-dind-patch.yaml kapp-controller-dns-patch.yaml config.toml tap-gui-secret.yaml kapp-controller-config-ca-overlay.yaml tce-tap-files/ 2>/dev/null
+  mv $cluster_name.yaml tap-values.yaml tap-gui-secret-patch.yaml tap-gui-dind-patch.yaml kapp-controller-dns-patch.yaml config.toml tap-gui-secret.yaml kapp-controller-config-ca-overlay.yaml $cluster_name-files/ 2>/dev/null
   echo "Your local TAP environment is ready!"
   if [[ $tap_profile == "full" ]]; then
     echo ""
